+++ include/gfx.h
@@ -23,6 +23,7 @@
 #include "core/gfx_timer.h"
 #include "core/gfx_obj.h"
 #include "widget/gfx_img.h"
+#include "widget/gfx_qrcode.h"
 #include "widget/gfx_label.h"
 #include "widget/gfx_anim.h"
 #include "widget/gfx_font_lvgl.h"