+++ include/widget/gfx_img.h
@@ -6,7 +6,6 @@
 
 #pragma once
 
-#include "core/gfx_types.h"
 #include "core/gfx_obj.h"
 
 #ifdef __cplusplus
@@ -26,7 +25,8 @@ extern "C" {
 
 /* Color format enumeration - simplified for public use */
 typedef enum {
-    GFX_COLOR_FORMAT_RGB565A8 = 0x0A,
+    GFX_COLOR_FORMAT_RGB565   = 0x04,  /**< RGB565 format without alpha channel */
+    GFX_COLOR_FORMAT_RGB565A8 = 0x0A,  /**< RGB565 format with separate alpha channel */
 } gfx_color_format_t;
 
 typedef struct {
@@ -73,7 +73,7 @@ gfx_obj_t *gfx_img_create(gfx_handle_t handle);
  * @param src Pointer to the image source data
  * @return Pointer to the object
  */
-gfx_obj_t *gfx_img_set_src(gfx_obj_t *obj, void *src);
+esp_err_t gfx_img_set_src(gfx_obj_t *obj, void *src);
 
 #ifdef __cplusplus
 }