+++ src/decoder/gfx_img_dec.c
@@ -8,10 +8,7 @@
 #include "esp_err.h"
 #include "esp_log.h"
 #include "esp_check.h"
-#include "core/gfx_core.h"
-#include "core/gfx_obj.h"
-#include "widget/gfx_img.h"
-#include "decoder/gfx_img_dec.h"
+#include "decoder/gfx_img_dec_priv.h"
 
 static const char *TAG = "gfx_img_decoder";