+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the ESP Emote GFX component will be documented in this file.
 
+## [2.0.0] - 2025-01-XX
+- Added partial refresh mode support
+- Added QR code widget (gfx_qrcode)
+
 ## [1.2.0] - 2025-09-0
 - use eaf as a lib